docker/src/requirements.txt
flask 
flask-login
flask-sqlalchemy 
psycopg2-binary
python-dotenv
Flask-WTF
email-validator